[execution-pool] Reorganize sender aware shuffler into directory
