+++ consensus/src/transaction_shuffler/sender_aware/config.rs
@@ -1,0 +1,6 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+#[derive(Clone, Debug)]
+pub struct Config {
+    pub conflict_window_size: usize,
+}
